pyproject.toml
[project]
name = "nonebot-plugin-shiro-personification"
version = "0.5.2"
description = "基于群聊与私聊上下文的人设回复插件，支持作息模拟、联网检索、风格学习、主动私聊、贴图、画像与 Agent 工具调用。"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "luojisama", email = "2534316454@qq.com" },
]
license = { text = "MIT" }
keywords = ["nonebot", "nonebot2", "plugin", "onebot", "personification", "chatbot"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Chat",
]

dependencies = [
    "nonebot2>=2.2.0",
    "nonebot-adapter-onebot>=2.4.0",
    "nonebot-plugin-apscheduler>=0.3.0",
    "nonebot-plugin-localstore>=0.6.0",
    "nonebot-plugin-htmlrender>=0.2.0",
    "openai>=1.0.0",
    "anthropic>=0.34.0",
    "filelock>=3.16.0",
    "watchdog>=4.0.0",
    "pydantic",
    "httpx",
    "Pillow",
    "PyYAML",
]

[project.urls]
"Homepage" = "https://github.com/luojisama/nonebot-plugin-personification"
"Bug Tracker" = "https://github.com/luojisama/nonebot-plugin-personification/issues"
"Repository" = "https://github.com/luojisama/nonebot-plugin-personification"

[tool.nonebot]
plugins = ["nonebot_plugin_personification"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
